editors - move some code over to browser/editor.ts
